requirements.txt
Django==2.2.12
elasticsearch-dsl==7.2.0
django-elasticsearch-dsl==7.1.1
djangorestframework==3.11.0
drf-yasg==1.17.1
python-Levenshtein==0.12.0
psycopg2-binary
djangorestframework-jwt==1.11.0
openpyxl==2.5.1  
googletrans==3.0.0
djangorestframework-simplejwt==4.4.0
django-request-logging==0.7.1
pyyaml==5.3.1
virtualenv
django-multiselectfield==0.1.12